Move L10n class from L10n subdir to Core (replacing old wrapper)
