+++ src/Console/PostUpdate.php
@@ -4,7 +4,7 @@ namespace Friendica\Console;
 
 use Friendica\App;
 use Friendica\Core\Config\IConfiguration;
-use Friendica\Core\L10n\L10n;
+use Friendica\Core\L10n;
 use Friendica\Core\Update;
 
 /**
