+++ src/Model/Storage/Database.php
@@ -6,7 +6,7 @@
 
 namespace Friendica\Model\Storage;
 
-use Friendica\Core\L10n\L10n;
+use Friendica\Core\L10n;
 use Psr\Log\LoggerInterface;
 use Friendica\Database\Database as DBA;
