+++ src/Module/Item/Ignore.php
@@ -4,7 +4,7 @@ namespace Friendica\Module\Item;
 
 use Friendica\App;
 use Friendica\BaseModule;
-use Friendica\Core\L10n\L10n;
+use Friendica\Core\L10n;
 use Friendica\Core\Session;
 use Friendica\Core\System;
 use Friendica\Database\Database;
