+++ src/Module/Register.php
@@ -7,7 +7,7 @@ use Friendica\Content\Text\BBCode;
 use Friendica\Core\Config;
 use Friendica\Core\Hook;
 use Friendica\Core\L10n;
-use Friendica\Core\L10n\L10n as L10nClass;
+use Friendica\Core\L10n as L10nClass;
 use Friendica\Core\Logger;
 use Friendica\Core\Renderer;
 use Friendica\Core\Worker;
