+++ tests/Util/SampleStorageBackend.php
@@ -6,7 +6,7 @@ use Friendica\App;
 use Friendica\Core\Hook;
 use Friendica\Model\Storage\IStorage;
 
-use Friendica\Core\L10n\L10n;
+use Friendica\Core\L10n;
 use Mockery\MockInterface;
 
 /**
