+++ tests/src/Core/L10n/L10nTest.php
@@ -2,7 +2,7 @@
 
 namespace src\Core\L10n;
 
-use Friendica\Core\L10n\L10n;
+use Friendica\Core\L10n;
 use Friendica\Test\MockedTest;
 
 class L10nTest extends MockedTest
